# skills/invoice-to-inventory/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Self Super Market API",
    "description": "REST API for managing products and inventory in Self Super Market",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Current environment"
    }
  ],
...
  "paths": {
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List all products",
        "description": "Retrieve a paginated list of products with optional search and category filtering",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search by name, barcode, or description",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a new product",
        "description": "Create a new product with the provided information",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or barcode already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Product ID"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "barcode": {
            "type": "string",
            "description": "Unique barcode"
          },
          "price": {
            "type": "string",
            "description": "Price in decimal format"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Product description"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "description": "Product category"
          },
          "stockQuantity": {
            "type": "integer",
            "description": "Stock quantity"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri",
            "description": "Product image URL"
          },
          "imagePublicId": {
            "type": "string",
            "nullable": true,
            "description": "Cloudinary public ID"
          },
          "isActive": {
            "type": "boolean",
            "description": "Product active status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "ProductCreate": {
        "type": "object",
        "required": [
          "name",
          "barcode"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Product name"
          },
          "barcode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "description": "Unique barcode"
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 1000000,
            "description": "Product price (optional)"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "nullable": true,
            "description": "Product description"
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "nullable": true,
            "description": "Product category"
          },
          "stockQuantity": {
            "type": "integer",
            "minimum": 0,
            "description": "Stock quantity"
          },
          "imageUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Product image URL"
          },
          "imagePublicId": {
            "type": "string",
            "nullable": true,
            "description": "Cloudinary public ID"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "description": "Items per page"
          },
          "total": {
            "type": "integer",
            "description": "Total number of items"
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  }
}